Follow-the-mouse focus was a bad idea
There are problems when running mplayer in full-screen - keyboard
controls won't work, depending on where the mouse was before the full
screen overlay opened. It was even impossible to stop the currently
playing movie other than switching to a text mode console, logging in
there, and then killing mplayer.

In every other regard, follow-the-mouse was quite useful. But
unfortunately, impaired movie playback is not an option for me.

Revert to click focus.