Separate quicklaunch entries from main menu config
This makes it easier to edit the quicklaunch settings without disturbing
the rest of the menu. It also avoids frequent file contents merges with
updates, because the main menu file will stay the same most of the time.

Also separate the browse/all and menu update entries for the same
reasons.