Use UNICODE arrow symbols in the window menu
Obviously, this required a font which actually provides those symbols.
The advantage is that arrows can show directions directly, without the
human reader needing to read lengthy text descriptions of the entries.

I even tried to use UNICODE symbols for fractions, but those are too
hard to read in my opinion because they are so narrow. Therefore using
normal text for fractions.

The available menu choices have also be augmented by adding sub-menus
for modifying position and size of windows independently.

Support for a width of 2/3rds has also been added; this is useful for
Web Browsers which need more horizontal space than half my screen.