Restructure Application menu
Hoist the submenu with the installed application out of a nested submenu
to the top-level menu.

In return, move the Workspaces menu from the top-level menu to the
nested submenu.

Practice showed that the Workspace menu is nearly never used, at least
as long as there is a taskbar present.