Get rid of "diameter"-symbol
The symbol is too hard to recognize and can be mistaken for a zero.

In order to maximize the visual distance between the grow/shring menu
entries, the factor is written first for growing. This allows to guess
the symbol even if it is tiny and hard to read based on the relative
position of the factor, because the denominator must always be right of
the symbol.

This also enforces the notion that numerator comes first and denominator
comes second, even if this is irrelevant in the case of a whole number.