Get rid of toolbar workspace controls
As it turns out, I rarely if ever use Fluxbox' workspaces. Therefore
remove the buttons for switching workspaces from the toolbar. (It is
still possible to use and switch workspaces via the menu.) This will
save some screen space.

However, it turned out there was a problem: When pressing the menu mouse
button somewhere in the icon bar area of the toolbar, frequently a
pop-up showing the full window title appears and remains on screen when
the menu is opened. This is ugly.

In order to provide an area somewhere in the left half of the toolbar
which does not suffer from this problem, move the buttons for switching
the current windows from the right side if the icon to its left side.

Also remove the button to make a window stick to all workspaces, which
has been used equally as rarely.